Move OSC 8 link demo into the primitives playground
Replaces the standalone `scripts/osc8-playground.tsx` + `osc8:playground`
command with a LinkText tab in the existing `pnpm try --playground` primitives
playground, matching the repo convention (demo under playground/demos/,
registered in PlaygroundApp).

Generated-By: PostHog Code
Task-Id: 6a298d39-7a8c-459a-869d-aee28a8dd037
